feat(crm): 完成商机阶段2联动开发

+++ Server/mitedtsm-module-crm/src/main/java/com/meession/etm/module/crm/dal/dataobject/business/CrmBusinessStatusDO.java
@@ -1,3 +1,4 @@
+// 23计科4班 黄金戈
 package com.meession.etm.module.crm.dal.dataobject.business;
 
 import com.meession.etm.framework.mybatis.core.dataobject.BaseDO;
@@ -5,6 +6,8 @@ import com.baomidou.mybatisplus.annotation.KeySequence;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.*;
+
+import java.math.BigDecimal;
 
 /**
  * CRM 商机状态 DO
@@ -40,7 +43,7 @@ public class CrmBusinessStatusDO extends BaseDO {
     /**
      * 赢单率，百分比
      */
-    private Integer percent;
+    private BigDecimal percent;
     /**
      * 排序
      */
